UserSkillsコンポーネント切り出し、コーディングスキルのみ反映
